--- lib/helpers/helpers.dart
@@ -1,5 +1,0 @@
-class PrimitiveWrapper {
-  var value;
-
-  PrimitiveWrapper(this.value);
-}
