+++ lib/helpers/primitive_wrapper.dart
@@ -1,0 +1,5 @@
+class PrimitiveWrapper<T> {
+  T value;
+
+  PrimitiveWrapper(this.value);
+}
